Local storage error try fix

+++ src/app/stores/auth/store.js
@@ -2,21 +2,27 @@
 import { create } from "zustand";
 import { persist, createJSONStorage } from 'zustand/middleware'
 
-
-// const auth = localStorage.getItem('auth') ? JSON.parse(localStorage.getItem('auth')) : null
-const auth = {}
-
-
-
 let initialAuthStore = {
-    ...auth
+    id: null,
+    name: null,
+    email: null,
+    role: null,
+    token: null,
+};
+
+if (typeof window !== 'undefined' && window.localStorage) {
+    const auth = localStorage.getItem('auth') ? JSON.parse(localStorage.getItem('auth')) : null;
+    initialAuthStore = {
+        ...initialAuthStore,
+        ...auth,
+    };
 }
 
 export const useAuthStore = create(persist(
     (set) => ({
         // initial state
         auth: initialAuthStore,
-        // Update auth store by loggin in
+        // Update auth store by logging in
         login: (payload) => {
             set((state) => ({
                 ...state.auth,
@@ -26,8 +32,7 @@ export const useAuthStore = create(persist(
                 role: payload.role,
                 token: payload.authToken
             }))
-    
-            return true
+            return true;
         },
         logout: () => {
             set((state) => ({
@@ -38,10 +43,12 @@ export const useAuthStore = create(persist(
                 role: null,
                 token: null
             }))
-            localStorage.clear();
+            if (typeof window !== 'undefined' && window.localStorage) {
+                localStorage.clear();
+            }
         }
     }), {
         name: 'auth',
         // storage: createJSONStorage(() => sessionStorage),
     }
-))
+));
